+++ Source/Krypton Components/Krypton.Ribbon/View Draw/ViewDrawRibbonCaptionArea.cs
@@ -1,4 +1,4 @@
-#region BSD License
+﻿#region BSD License
 /*
  * 
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
@@ -343,6 +343,24 @@ internal class ViewDrawRibbonCaptionArea : ViewDrawDocker
 	public void PerformFormChromeCheck() =>
 		// Update decision about integrating or providing caption functionality
 		OnFormChromeCheck(null, EventArgs.Empty);
+	#endregion
+
+	#region ApplyPaletteChanged
+	/// <summary>
+	/// Re-evaluate QAT, Office 2007 app-button, form-icon, and injected caption chrome after a palette change.
+	/// </summary>
+	/// <remarks>
+	/// <see cref="KryptonRibbon.RibbonShape"/> comes from the palette. Call this from
+	/// <see cref="KryptonRibbon.OnPaletteChanged"/> so caption chrome updates on a global theme
+	/// swap without waiting for a resize or form-chrome event (#3859, #4061).
+	/// </remarks>
+	public void ApplyPaletteChanged()
+	{
+		UpdateQAT();
+		AppButtonVisibleChanged();
+		PerformFormChromeCheck();
+		RedrawCustomChrome(true);
+	}
 	#endregion
 
 	#region DoesClientMouseDownEndAllTracking